Any Day Commit 6
Continued work on scheduling points to any day. Got the basis for the
algorithm implemented. Using 2 random numbers within a range of how many
days are to be scheduled vs how many are not for a certain point. The
greater random number determines whether it is routed that day. Still
needs implemented fully and tested.